cosim を EIP 32bit化に追随させる
CPU照合 (CI) だけが捕まえたビルド漏れ。cosim は default-members の外に
居るので、手元の cargo build --all-targets も clippy --workspace
--exclude rustx86-cosim も、全部この crate を素通りしていた。
Unicorn のCビルドが重いから除外している以上、ここを見張れるのは
paths で発火する CPU照合 ワークフローだけ — 設計どおりの網に掛かった。

State::ip は u16 のまま、渡す側を切り詰める。MODE_16 の突き合わせでは
IPは16bitが正しい。MODE_32 対応のときに器ごと広げる。

手元でも cargo test -p rustx86-cosim を実際に回した: 15件全通過
(DAAの総当たり含む、112秒)。

+++ cosim/src/lib.rs
@@ -105,7 +105,7 @@ pub fn run_ours(tc: &TestCase) -> State {
         regs: std::array::from_fn(|i| m.cpu.regs[i] as u16),
         sregs: std::array::from_fn(|i| m.cpu.sregs[i]),
         flags: m.cpu.flags as u16 & FLAG_MASK_ALL,
-        ip: m.cpu.ip,
+        ip: m.cpu.ip as u16,
         data,
         stack,
     }
